https://bugzilla.redhat.com/show_bug.cgi?id=452866
The _netdev option is used in /etc/fstab to mark devices that depend on the
network for operation. This causes the initscripts to defer mounting file
systems located on these devices until the network is brought up and the netfs
initscript runs.

Device nodes for these devices (nbd, gnbd, iscsi etc.) are created by udev once
the appropriate subsystem is started.

Processing of the udev backlog may not have completed by the time the netfs
script runs causing failures to occur when attempting to mount file systems.
This can happen when there are a large number of devices or when device
initialisation is slow.

This is fixed by adding a udevsettle to the start() function in netfs.