Add a address check capability for IPv4 and IPv6 addresses. (#464268)
Add the -c, --check argument which enables address validation.  The
ipcalc program will exit with 0 if the address is valid or 1 if it
is invalid, for the specified family.

Add the -4, --ipv4 argument to specify the IPv4 address family.

Add the -6, --ipv6 argument to specify the IPv6 address family.

Ensure that the IPv6 family argument is invalid with the IPv4-only
show arguments (broadcast, network, netmask, prefix).