Don't directly execute bash for single-user mode, as that has SELinux issues. (#540216, #584443, #585053)
If upstart directly runs bash for single-user mode, there's no way for
SELinux to set the proper context (to distinguish it from other init scripts).
Inserting a helper binary in the way allows for this.

While we're changing this, we make it configurable so it's easily possible
for people to use sulogin if they so desire.