systemctl: honour $SYSTEMCTL_IGNORE_DEPENDENCIES
Here's a patch for initscripts:

If SYSTEMCTL_IGNORE_DEPENDENCIES is set, pass --ignore-dependencies to
systemctl. This allows programs which invoke /sbin/service to get more
SysV-like behaviour, i.e. where starting a service does not pull in any
dependencies and honours no ordering.

By default we continue to honour the dependencies since this is usually
safe and allows other software to hook into all services regardless by
who they are started.

This is the result of a discussion with the IPA people, Simo Sorce in
particular.