fix sysctl calls with VLAN (mostly breaking IPv6)
[Bill Nottingham <notting@redhat.com>: updated patch]

sysctl uses "." as the path delimiter. VLAN uses "." as the id
delimiter. initscripts does absolutely nothing to resolve this conflict.
The end result is that all sysctl calls on VLAN if:s will just fail
(silently for extra bonus as the scripts pipe to /dev/null).

With legacy IP this isn't a big deal as the sysctl use is very sparse.
For IPv6 however, the shit really hits the fan as you can't even
disable address autoconfiguration without sysctl.

https://bugzilla.redhat.com/show_bug.cgi?id=665601