service: add support for legacy custom actions
This adds support for legacy initscript actions such as
  service iptables save

Packages wishing to provide legacy actions should install programs (probably
scripts, but could be binaries) named
  /usr/libexec/initscripts/legacy-actions/<service>/<action>

It should be emphasized that this feature is meant to be used only in packages
where compatibility of custom actions with previous releases is desired.
Adding of entirely new actions is discouraged. Useful tools should be
shipped by upstreams as programs in /usr/bin, not in distro-specific hacks.

[ Modifications from original : change from /usr/lib to /usr/libexec - notting ]

https://bugzilla.redhat.com/show_bug.cgi?id=796663