Optionally do not send hostname with dhclient -H
When using DHCP and invoking dhclient, ifup-eth always passes the -H
flag which provides a hostname to the DHCP server.  On systems with
multiple interfaces it is sometimes required for only one interface to
send a hostname.

This patch adds the DHCP_SEND_HOSTNAME and the DHCP6_SEND_HOSTNAME
options which, when set to "no" or "off", remove the -H from the
dhclient command.  These are added to the user's ifcfg-$INTERFACE file.