ifup/ifdown: use alternatives system
This installs the real scripts into /etc/sysconfig/network-scripts while
leaving the management of the launchers in /usr/sbin to the alternatives
systemd.

This allows coexistence with alternative implementations of ifup and
ifdown, notably NetworkManager.

A priority of 90 is chosen rather arbitrarily, NetworkManager will use a
lower one (50) so that our implementation takes precedence over
NetworkManaager in case both are installed. In an unlikely event the
user will want to switch to NetworkManager implementations when both are
installed they can override the default with:

  # update-alternatives --config ifup