Add optional 'dev' keyword
Fix the problem when the device name could be interpreted as an iproute2 keyword.
For example, for a bridge slave named "a" the iproute2 would treat
the name as a prefix of keyword "address" and the network-scripts
would fail to set the bridge master.

Resolves: rhbz #1859785

(cherry picked from commit 100ed46b483b50059b3be796bafaf7b8f2b99220)