ifup-routes: Use `ip route repace` to avoid race
This should resolve the scenario when the link is brought
up prior to disabling accept_ra.

This only happens when both IPv4 and IPv6 address is on
the interface, because network-scripts does IPv4 first
and brings up the interface.

Then it does IPv6 which disables the RA's, at that point
the node has already learned the route from RA and setting
a default route fails.

Using `ip route replace` we ensure if the above scenario
happens we end-up with the correct default ipv6 route.

Huge thanks to Harald who debugged this issue and prepared a patch!

Related: #2034799

Co-authored-by: hjensas <hjensas@redhat.com>