netconsole: source sysconfig/network only if it exists
Although the comment said "check if network is up",
the code only sourced the configuration file, but did not fail
if it did not exist and happily continued.

Also on current versions of Fedora, the sysconfig/network
does not exist by default, so you see a warning message
when you start the service.

So let's just check if the file exists and only source it
if it does (for backwards compatibility).