network-functions: Fix a bug in is_nm_handling()
Currently the "^" anchor is applied to the "connected" case and does not
apply to the "connecting" case in the grep due to the way the brackets
() are placed. This can cause the grep statement to incorrectly return a
match.

For ex: When trying to check for eth0 using is_nm_handling(),
"bondeth0:connecting" would also match the grep and return incorrectly.

Signed-off-by: Rohit Nair <rohitnair2297@gmail.com>